requirements.txt
pyyaml
requests
python-gyazo
